requirements.txt
# Django Core
Django>=5.0,<6.0
djangorestframework>=3.14,<4.0
django-filter>=23.5
django-cors-headers>=4.3

# Authentication (allauth)
django-allauth[mfa,socialaccount]>=65.0
fido2>=2.1

# Image Processing
Pillow>=10.0

# HTML Sanitization
bleach>=6.0

# QR Code Generation
qrcode[pil]>=7.4

# Date/Time Utilities
python-dateutil>=2.8

# Async Tasks
celery>=5.3
redis>=5.0
django-celery-beat>=2.5

# Environment Configuration
django-environ>=0.11

# Testing
pytest>=7.4
pytest-django>=4.7
factory-boy>=3.3
pytest-cov>=4.1
freezegun>=1.2

# Code Quality
flake8>=6.1
black>=23.12
isort>=5.13

# API Documentation
drf-spectacular>=0.27

# Payments
stripe>=7.0

# PDF Generation
xhtml2pdf>=0.2.11
